parte 10

+++ desktop/src/com/mygdx/game/desktop/DesktopLauncher.kt
@@ -8,8 +8,8 @@ object DesktopLauncher {
     @JvmStatic
     fun main(arg: Array<String>) {
         val config = LwjglApplicationConfiguration()
-        config.width = FlappyGame().WIDTH
-        config.height = FlappyGame().HEIGHT
+        config.width = FlappyGame().WIDTH.toInt()
+        config.height = FlappyGame().HEIGHT.toInt()
         config.title = FlappyGame().TITLE
         LwjglApplication(FlappyGame(), config)
     }
